full-package/requirements.txt
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
umap-learn>=0.5.0
seaborn>=0.11.0
click>=8.0.0
pandas>=1.3.0
torch>=1.9.0
optuna>=3.0.0
python-dotenv>=0.19.0
openai>=0.27.0
llama-cpp-python>=0.2.0
streamlit>=1.25.0
faiss-cpu>=1.7.0
hdbscan>=0.8.0 